Initialize repo with Modelica code
The library has been initialized with code from
Modelica.Electrical and Modelica.Blocks. Since
a lot of code is going to be borrowed from these
two libraries, this seems to be a good starting
point. In the future, models can be deleted
once it's clear that they won't be used.