Adds Scripts folder
This folder has been added to contain scripts used
with the library. The scripts included in this commit
are in MATLAB and Python and are aimed at postprocessing
Dymola results. The Python script is the most successful
one right now. In the future, this folder could contain
scripts for running simulations in Dymola also.