Add basic switch models
This commit includes a new package, Electrical.PowerConverters
with three models: a basic ideal bidirectional switch, a two
switch leg and an H-bridge. Each is achieved by composition
of the previous.

This commit also includes a small Not block needed for the
generation of the bottom firing signal (as the complement of
the top firing signal).