Change IdealBuckOpen example to use SignalPWM
As mentioned in a previous commit the IdealBuckOpen example wasn't
really using any original blocks or models from PVlib. In this
commit that is changed by using the SignalPWM block, which
additionally allows changing the duty cycle during the simulation,
better showcasing the library and providing a better examples use.