Add icons for basic original components
This commit adds icons for the different components developed, which
include Not and Less, the three switch models and the PWM block. The
commit also includes quite a lot of modifications introduced by
Dymola which I'm not sure I can avoid, so there they go. It's messy
but better than a kick in the teeth.