Update to MSL 2.2.1 and massive cleanup
After indication from my supervisor, I moved on to 2.2.1 in Dymola
6.1 and migrated the library. This involved a massive update which
I decided to commit in one go.

The numerous amount of borrowed blocks and models from MSL has
been eliminated from the library since it feels way better to
reuse code by referencing instead of copying. This results in a
much leaner library which will be easier to understand, mantain
and develop.

The library is still lacking in good documentation through
annotations. Some of the info presented in the ones that are there
is also incorrect. This will be solved in future commits.