Change name and fix HBridge models
It seems like a good idea to drop the Ideal prefix since I'm not
sure what I mean by that anyway and it's cluttering the names.
Additionally included switched and averaged sufixes now that we
have two variants.

With all the cleaning up and name changing, HBridgeAveraged had
a couple of connects using the old name of the duty cycle input
port. This patch fixes that.