Split models into stand-alone files
I'm placing each model into its own file. I'm not sure how good of an
idea this is, but it's convenient to be able to reference the code in
LaTeX using listings. I have the impression that this splitting is not
normal practice, but I have seen it done sometimes. Depending on how
it works out, I might put things back together in the future.