Modify BuckOpen example to use replaceable switch
It works! For the moment I just wanted to try out the replaceable
feature and it seems quite handy. By choosing CCM-DCM1 with Le=1e-6
and fs=1e5, the results of the average model are much closer to the
switching model. This will be included in the documentation of the
class (with graphs of different results) in a future commit.