Add averaged CPM models
They are still not validated, but are included in this commit in order
to keep things tidy. I'm creating the capstone bidirectional buck
boost converter example to validate the CPM blocks, but I'm finding I
might want to reorganize things a little, which will probably come in
the next commit.