Add Electrical.Assemblies package and CPM example
After some thought, it seems like the right way to go is to add an
assemblies package inside Electrical. This will hold arbitrary
converter models and maybe some other assemblies (i.e. a battery +
converter system).

The two existing converter models were moved under this package and a
new converter, the bidirectional buck boost, was added. This last
converter is used in a preliminary example in CPM open loop as a
starting point to the capstone example.