Improve SignalPWM implementation
This version has been adapted from PowerSystems. It has a much better
icon, and more functionality. Compared to the implementation from
PowerSystems, this one doesn't have the constant duty option, the
complement of the firing signal can be enabled/disabled and when it's
enabled, a dead time can be specified (ignored when going back to only
one firing signal).

This commit includes modifications to several models that depended on
this block.