Remove zero order hold from SignalPWM
It seems like the model is better, more faithful, with the zero order
hold (just as the PWM from PowerConverters), but it creates a
discrepancy between the switching and averaged results which is bit
ugly. For the moment, I won't be including this zero order hold for
the sake of getting congruent results. If I'm able to understand and
explain this effect better, I'll consider including it.