Add switching CPM block
As a first implementation, this seems to do the job quite well. Taken
from Power Electronics Specialization at Coursera. A preliminary
validation model is also added. The names of the different CPM blocks
is also standardised.

This block will enable modelling of switching CPM converters which
will enable the validation of the averaged models. Notably, this can
be used in the switching version of the bidirectional buck boost
converter used in the USB battery example.