Add Control.Assemblies package
The controller blocks have been moved under this package and the
description of the Electrical.Assemblies package has been shortened
and standardised with this one.

The distinction between a component and an assembly model is a bit
fuzzy. For the moment, this structure for the library makes sense
although it might seem a bit arbitrary.