Delete Ideal2LevelLeg
This model included a deadtime and routing of the firing signals. When
used to build the HBridgeSwitched converter, this created questions as
to where to include the dead time block, and also required several
layers of parametrization and propagation. It seems a lot of
complexity for the seemingly reduced applicability of the
assembly. The HBridgeSwitched converter can be constructed from
current bidirectional switches without making a mess. Seems cleaner.