Standardize converter interfaces and icons
I discovered that HBridge converters were not inheriting from
TwoPortConverter, which was not the intention. Also, icons were
normalized for all three converters. Two questions remaining, for
future commits:
+ Both the HBridge converters and the buck boost are using the same
  topology (H-bridge).
+ The topology can be used in both AC/DC and DC/DC applications, so
  it's a bit arbitrary to include the AC and DC symbols in the
  icon.

It seems like it would be a good idea to find a list of common
topologies somewhere and provide models for them with appropriate use
of composition and inheritance. For future commits.