Fix inverter current controller and examples
I reviewed the controller and associated examples and they seem to be
working now. I also improved the interface and icon of the controller
and cleared up the diagrams of the examples (swapping sensor blocks
for tidier real expression blocks).

I still get an error when trying to simulate Inverter1phClosedSynch
with step blocks on the dq setpoint inputs. The translation fails
because it cannot differentiate the higher index DAE. Using constant
blocks gets around the problem, but these models will probably need
more work.