Refactor CPMBidirectionalBuckBoost
I'm having some problems with the simulation of the
USBBatteryConverter example and am doing some cleaning and refactoring
in the process. This commit mainly includes a new
BidirectionalBuckBoost to which the CPM control blocks are added to
produce the original converter. The problems have to do with something
changing too fast (in the buck to boost mode change) or maybe creating
a non-solvable situation. Will investigate further.