Add disable functionality to CPM_CCM
This block includes a non-linear equation that needs to be
solved. When the block is being used (it's d output being used), it
normally has inputs that make sense. In some cases, the output might
be ignored for some time intervals, but the solver still needs to
solve the equation. In these cases, it makes sense to be able to
disable de block.

It might be convenient to be able to make this input conditional and
it will probably also make sense to include it in the CCM DCM
implementation of the CPM modulator, but this commit is a starting point.