Add switched switch network models
I've had this idea to create switched versions of the average switch
network models. I've had to change some of the partial classes from
which different electrical models were inheriting but it's paid
off. Now, a user can modify an instance of the HBridgeAveraged model
to make it behave like the switched version. Amazing! Now it doesn't
make sense to have average and switched versions of the converter,
since it can be changed in this way on the diagram. Clean-up will come
in subsequent commits.