Add preliminary scripts
As a starting point, these scripts have enabled me to explore how I
can use them. The goal is to provide scripts to run all the
examples (for better user experience), as well as have scripts to run
all of the validation examples, and hopefully create some kind of
report. I'm also hoping to create scripts to generate SVG files of all
icons and diagrams as well as PNG of plots of examples. My experience
with these last ones is that SVG is not very well supported (i.e. I
got an image with a black background)