Change SignalPWM icon
The icon was copied from the same block available in MSL. Since the
functionality is a bit different, it seems a good idea to change the
icon in order to avoid confusion. Also, the icon is now more coherent
with the rest of the icons of the library. The name will also
change, for the same reasons, in a future commit (since it will
require saving all of the dependent models.