Improve switched models usage
The name of HBridgeAveraged is changed to HBridge since it can now be
instantiated and modified to work as a switched or an averaged
model. For the moment, the HBridgeSwitched model is still kept, since
I'm still exploring if this different approach is superior.

I tested all of the affected application examples and it seems like it
could be correct. Some strange results happen for the last, more
complex, examples (PVInverter...). I created a bidirectional version
of the switched switch network (i.e. with an anti-parallel diode) but
it creates the same results as the synchronous on which it's based.