Complete Examples documentation
The documentation is not very detailed and USBBatteryConverter is
still undocumented, but this commit provides an improvement. All
examples, including the ones in Application and Verification should
now have a StopTime value, which will mean they will be run for
regression tests. The reference data for regression tests needs to be
updated, though.