Fix CPMBidirectionalBuckBoost mode change
In previous releases, USBBatteryConverter could not be simulated past
10 seconds, when a buck to boost mode change of the converter was
scheduled, because the solver couldn't find a solution at the event.

This commit adds an on delay block to the enable signals of both the
buck and boost CPMs, to alleviate this situation. This doesn't seem
like the best solution but will do for the moment. A better solution
might be to program the CPM blocks to provide a more event-friendly
solution.