Add white rectangle for better SVG-PDF export
The way I'm producing diagrams for the LaTeX thesis document is by
exporting SVG diagrams with Dymola and using command-line Inkscape to
export that into PDF, which can then be included in the LaTeX
document. Some part of the process crops images a bit too tight and
some labels and connections are sometimes lost. As a hack, including a
thin white rectangle around the blocks solves this. Not ideal, but I
can always remove this hack when I find a cleaner strategy.